# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sidequestd API",
    "description": "Shared API for the Sidequestd web and mobile clients.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/games": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Browse the catalog",
        "description": "Filter by genre and/or platform (SPEC \u00a76.5 Browse).\n\nRepeated values within a facet are OR-ed and the facets are AND-ed, so\n`?genre=indie&genre=puzzle&platform=nintendo-switch` reads as \"indie or\npuzzle, on Switch\".",
        "operationId": "browse_games_api_v1_games_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Genre slug. Repeat for any-of.",
              "title": "Genre"
            },
            "description": "Genre slug. Repeat for any-of."
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Platform slug. Repeat for any-of.",
              "title": "Platform"
            },
            "description": "Platform slug. Repeat for any-of."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GameSort",
              "description": "Result ordering",
              "default": "title"
            },
            "description": "Result ordering"
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TrendingWindow",
              "description": "Rolling window the ranking covers",
              "default": "7d"
            },
            "description": "Rolling window the ranking covers"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "From a previous page",
              "title": "Cursor"
            },
            "description": "From a previous page"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Results per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Results per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_GameSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/games/trending": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Trending games for a rolling window",
        "description": "Reads the ranking materialised by `python -m app.cli.trending`.\n\nEmpty until that has run against some activity \u2014 the endpoint does not fall\nback to an arbitrary list, because a made-up \"trending\" is worse than none.",
        "operationId": "trending_games_api_v1_games_trending_get",
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TrendingWindow",
              "description": "Rolling window the ranking covers",
              "default": "7d"
            },
            "description": "Rolling window the ranking covers"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Results per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Results per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrendingGame"
                  },
                  "title": "Response Trending Games Api V1 Games Trending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/discover": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "The Discover tab's sections in one call",
        "description": "One round trip for the whole tab (SPEC \u00a76.5).\n\n`recommended` is personal to the caller: their favourite genres and the\ngenres of what they rated well, blended with what the people who rate like\nthem rated well, minus everything they have already reviewed, listed or\npinned. A caller we know nothing about \u2014 signed out, or signed in and new \u2014\ngets what is trending and what has always been popular instead, so the\nsection is never an error and never a lie about being tailored.",
        "operationId": "discover_api_v1_games_discover_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TrendingWindow",
              "description": "Rolling window the ranking covers",
              "default": "7d"
            },
            "description": "Rolling window the ranking covers"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Full detail for one game",
        "description": "Declared last on purpose: `/games/trending` and friends are literal paths\nthat would otherwise be swallowed by this one.",
        "operationId": "read_game_api_v1_games__game_id__get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Catalog id",
              "title": "Game Id"
            },
            "description": "Catalog id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "404": {
            "description": "No such game"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/games": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search games by title",
        "description": "Results carry cover art, release year and platforms, which is what SPEC\n\u00a76.6 says a result card shows.",
        "operationId": "search_games_api_v1_search_games_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "What the user typed",
              "title": "Q"
            },
            "description": "What the user typed"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "From a previous page",
              "title": "Cursor"
            },
            "description": "From a previous page"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Results per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Results per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_GameSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/users": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search users by handle or display name",
        "description": "Private accounts appear as shells (SPEC \u00a76.6, \u00a76.7).\n\nThey have to be findable \u2014 otherwise nobody could send a follow request \u2014\nbut `review_count` comes back null unless the viewer is the account itself\nor an approved follower.",
        "operationId": "search_users_api_v1_search_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "What the user typed",
              "title": "Q"
            },
            "description": "What the user typed"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "From a previous page",
              "title": "Cursor"
            },
            "description": "From a previous page"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Results per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Results per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_UserSearchResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CursorPage_GameSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as ?cursor= for the next page; null when exhausted."
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CursorPage[GameSummary]"
      },
...
      "CursorPage_UserSearchResult_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserSearchResult"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as ?cursor= for the next page; null when exhausted."
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CursorPage[UserSearchResult]"
      },
      "DiscoverResponse": {
        "properties": {
          "trending": {
            "items": {
              "$ref": "#/components/schemas/TrendingGame"
            },
            "type": "array",
            "title": "Trending"
          },
          "new_releases": {
            "items": {
              "$ref": "#/components/schemas/GameSummary"
            },
            "type": "array",
            "title": "New Releases"
          },
          "recommended": {
            "items": {
              "$ref": "#/components/schemas/GameSummary"
            },
            "type": "array",
            "title": "Recommended",
            "description": "Personalised for the caller, or trending and popular all-time for a caller with no ratings, favorites or session."
          },
          "genres": {
            "items": {
              "$ref": "#/components/schemas/GenreRef"
            },
            "type": "array",
            "title": "Genres"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/PlatformRef"
            },
            "type": "array",
            "title": "Platforms"
          }
        },
        "type": "object",
        "required": [
          "trending",
          "new_releases",
          "recommended",
          "genres",
          "platforms"
        ],
        "title": "DiscoverResponse",
        "description": "The Discover tab's first paint (SPEC \u00a76.5).\n\nEvery section is a plain list of cards, `recommended` included: the ranking\nthat produced it is the server's business, and a score on the wire is a\nnumber a client would be tempted to show. See `app.services.recommendations`\nfor how the order is arrived at."
      },
...
      "GameDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/PlatformRef"
            },
            "type": "array",
            "title": "Platforms"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "genres": {
            "items": {
              "$ref": "#/components/schemas/GenreRef"
            },
            "type": "array",
            "title": "Genres"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "external_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "release_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Year",
            "description": "Precomputed so every client renders the same year for a given date.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "cover_url",
          "release_date",
          "platforms",
          "summary",
          "genres",
          "external_id",
          "external_source",
          "created_at",
          "release_year"
        ],
        "title": "GameDetail",
        "description": "Everything the Game Detail screen needs (SPEC \u00a75)."
      },
      "GameSort": {
        "type": "string",
        "enum": [
          "title",
          "release_date",
          "trending"
        ],
        "title": "GameSort",
        "description": "Orderings offered by `GET /games` (SPEC \u00a76.5 Browse)."
      },
      "GameSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/PlatformRef"
            },
            "type": "array",
            "title": "Platforms"
          },
          "release_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Year",
            "description": "Precomputed so every client renders the same year for a given date.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "cover_url",
          "release_date",
          "platforms",
          "release_year"
        ],
        "title": "GameSummary",
        "description": "The card shape used by browse and search results.\n\nDeliberately narrow: SPEC \u00a76.6 asks for cover art, release year and platform\non a result, and shipping the long `summary` text in a 20-item page would be\nmost of the payload for something no card renders."
      },
      "GenreRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "GenreRef"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlatformRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "PlatformRef"
      },
...
      "TrendingGame": {
        "properties": {
          "game": {
            "$ref": "#/components/schemas/GameSummary"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "window": {
            "$ref": "#/components/schemas/TrendingWindow",
            "description": "Rolling window the score was computed over"
          }
        },
        "type": "object",
        "required": [
          "game",
          "score",
          "window"
        ],
        "title": "TrendingGame",
        "description": "A trending entry: the card plus the score that ranked it (SPEC \u00a76.11)."
      },
      "TrendingWindow": {
        "type": "string",
        "enum": [
          "24h",
          "7d",
          "30d"
        ],
        "title": "TrendingWindow",
        "description": "Rolling windows the trending worker computes (SPEC \u00a76.11).\n\nClosed set rather than free text so a typo can't quietly return an empty\nranking; the durations live in `app.services.trending.TRENDING_WINDOWS`."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /auth/login",
        "scheme": "bearer"
      }
    }
  }
}